Add files via upload
These two codes combined make the bare bone basic essentials to play a classic game of checkers. The players will be able to input their name, move pieces according to the rules of checkers, all while the code keeps track of the score that delivers it back to each player until the victor is decided.